Switch Docker base to jlesage/docker-baseimage-gui

+++ src/modules/minimalkioskos/filesystem/home/pi/scripts/startup.sh
@@ -6,14 +6,16 @@ export DISPLAY
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 # Clear Chromium config and cache, they tend to corrupt themselves, causing Chromium to segfault
-rm -rf "$HOME/.config/chromium"
-rm -rf "$HOME/.cache/chromium"
+if [ -z "$RUNNING_IN_DOCKER" ]; then
+    rm -rf "$HOME/.config/chromium"
+    rm -rf "$HOME/.cache/chromium"
+fi
 
 # Autohide mouse when inactive
 unclutter &
 
 # Give pi account a complex random password
-if [ -e "/boot/autosecure" ]
+if [ -z "$RUNNING_IN_DOCKER" ] && [ -e "/boot/autosecure" ]
 then
     NEWPW="$(openssl rand -base64 32 | tr -d 'EOF')"
     passwd <<EOF
@@ -24,12 +26,21 @@ EOF
 fi
 
 # Start Python-based controller to ensure reloads on load failures
-python3 "$DIR/chromium_controller.py" "$(head -n 1 /boot/mutesound.txt)" &
+if [ -n "$RUNNING_IN_DOCKER" ]; then
+    python3 "$DIR/chromium_controller.py" "$(head -n 1 /config/mutesound.txt)" &
+else
+    python3 "$DIR/chromium_controller.py" "$(head -n 1 /boot/mutesound.txt)" &
+fi
+
+BROWSER="$(command -v chromium-browser)"
+[ -z "$BROWSER" ] && BROWSER="$(command -v chromium)"
+
 
 # Start Chromium
-while true; do
-    chromium-browser --kiosk --touch-events=enabled --disable-pinch --noerrdialogs --disable-session-crashed-bubble --start-fullscreen --remote-debugging-port=9222 --app="file:///boot/placeholder.html"
-    if [ -n "$RUNNING_IN_DOCKER" ]; then
-        exit "$?"
-    fi
-done
+if [ -n "$RUNNING_IN_DOCKER" ]; then
+    $BROWSER --kiosk --touch-events=enabled --disable-pinch --noerrdialogs --disable-session-crashed-bubble --start-fullscreen --remote-debugging-port=9222 --app="file:///config/placeholder.html"
+else
+    while true; do
+        $BROWSER --kiosk --touch-events=enabled --disable-pinch --noerrdialogs --disable-session-crashed-bubble --start-fullscreen --remote-debugging-port=9222 --app="file:///boot/placeholder.html"
+    done
+fi
